# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "G-STK Back-End OpenApi  Documentation",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/entreprises": {
            "get": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Get list of entreprises",
                "description": "Returns list of entreprises",
                "operationId": "getEntreprisesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntrepriseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Store new entreprise",
                "description": "Returns entreprise data",
                "operationId": "storeEntreprise",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreEntrepriseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Entreprise"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "api/entreprises/{id}": {
            "get": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Get entreprise information",
                "description": "Returns entreprise data",
                "operationId": "getEntrepriseById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Entreprise id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Entreprise"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Update existing entreprise",
                "description": "Returns updated entreprise data",
                "operationId": "updateEntreprise",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Entreprise id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEntrepriseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Entreprise"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Delete existing entreprise",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteEntreprise",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Entreprise id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Entreprise": {
                "title": "Entreprise",
                "description": "Entreprise model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "public": {
                        "title": "Public",
                        "type": "boolean"
                    },
                    "website_link": {
                        "title": "Websitelink",
                        "type": "string",
                        "example": "http://schimmel.com/"
                    },
                    "fichier_id": {
                        "title": "fichier_id",
                        "description": "For logo",
                        "type": "integer"
                    },
                    "presentation": {
                        "title": "Presentation",
                        "type": "string"
                    },
                    "module_entreprise_id": {
                        "title": "Moduleentrepriseid",
                        "description": "Module Entreprise Id",
                        "type": "integer"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "module_id": {
                        "title": "Module's id",
                        "description": "Id of module",
                        "type": "integer"
                    },
                    "user_id": {
                        "title": "User's id",
                        "description": "Id of User",
                        "type": "integer"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Entreprise"
                }
            },
...
            "StoreEntrepriseRequest": {
                "title": "Store Entreprise request",
                "description": "Store Entreprise request body data",
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "public": {
                        "title": "Public",
                        "type": "boolean"
                    },
                    "website_link": {
                        "title": "Websitelink",
                        "type": "string",
                        "example": "http://schimmel.com/"
                    },
                    "fichier_id": {
                        "title": "fichier_id",
                        "description": "For logo",
                        "type": "integer"
                    },
                    "presentation": {
                        "title": "Presentation",
                        "type": "string"
                    },
                    "module_entreprise_id": {
                        "title": "Moduleentrepriseid",
                        "description": "Module Entreprise Id",
                        "type": "integer"
                    },
                    "module_id": {
                        "title": "Module's id",
                        "description": "Id of module",
                        "type": "integer"
                    },
                    "user_id": {
                        "title": "User's id",
                        "description": "Id of User",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "UpdateEntrepriseRequest": {
                "title": "Update Entreprise request",
                "description": "Update Entreprise request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "public": {
                        "title": "Public",
                        "type": "boolean"
                    },
                    "website_link": {
                        "title": "Websitelink",
                        "type": "string",
                        "example": "http://schimmel.com/"
                    },
                    "fichier_id": {
                        "title": "fichier_id",
                        "description": "For logo",
                        "type": "integer"
                    },
                    "presentation": {
                        "title": "Presentation",
                        "type": "string"
                    },
                    "module_entreprise_id": {
                        "title": "Moduleentrepriseid",
                        "description": "Module Entreprise Id",
                        "type": "integer"
                    },
                    "module_id": {
                        "title": "Module's id",
                        "description": "Id of module",
                        "type": "integer"
                    },
                    "user_id": {
                        "title": "User's id",
                        "description": "Id of User",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "EntrepriseResource": {
                "title": "Entreprise Resource",
                "description": "Entreprise resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Entreprise"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "EntrepriseResource"
                }
            },
...
        }
    },
    "security": [
        []
    ]
}